fixed a game breaking bug
resolved #116

The cause was found to be the spectator mode once you die. As soon as your camera starts "bouncing" between players, the myID(global) variable is set to whatever player you are currently spectating

upon rejoining the lobby the myID variable is now resent by the server

also resolved a smaller issue with the "start game now vs. AI" button appearing when it shouldnt have